requirements.txt
streamlit
gTTS
langchain
langchain-community
langchain-google-genai
langchain-groq
google-generativeai
chromadb
PyPDF2
